SDGToggle 관련 Style 및 Colors 클래스 별도 파일로 분리
