+++ sdg/src/main/java/com/shopl/sdg/component/toggle/style/SDGToggleColors.kt
@@ -1,0 +1,14 @@
+package com.shopl.sdg.component.toggle.style
+
+import androidx.compose.ui.graphics.Color
+
+internal data class SDGToggleColors(
+    val onTrackColor: Color,
+    val onThumbColor: Color,
+    val offTrackColor: Color,
+    val offThumbColor: Color,
+    val disabledOnTrackColor: Color,
+    val disabledOnThumbColor: Color,
+    val disabledOffTrackColor: Color,
+    val disabledOffThumbColor: Color,
+)
